requirements-gcp.txt
flask==1.0.2
google-cloud-storage==1.13.2